+++ app/src/androidTest/java/com/leonis/android/adhafera/lib/HTTPClientTest.java
@@ -18,6 +18,7 @@ import java.util.HashMap;
 
 import static android.support.test.InstrumentationRegistry.getContext;
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.fail;
 import static org.mockito.Mockito.doNothing;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
@@ -43,8 +44,20 @@ public class HTTPClientTest {
     @Test
     public void testRegisterPayment_OK() {
         String[] inputs = {"2015-01-01", "for http client test", "test", "100", "expense"};
-        HTTPClient httpClient = new HTTPClient(getContext(), inputs);
-        setupMock(httpClient, 201, "{\"date\":\"2015-01-01\",\"content\":\"for http client test\",\"category\":\"test\",\"price\":\"100\",\"payment_type\":\"expense\"}");
+        HTTPClient httpClient = new HTTPClient(getContext());
+        httpClient.createPayment(inputs);
+        JSONObject payment = new JSONObject();
+        try {
+            payment.put("date", inputs[0]);
+            payment.put("content", inputs[1]);
+            payment.put("category", inputs[2]);
+            payment.put("price", inputs[3]);
+            payment.put("payment_type", inputs[4]);
+        } catch (JSONException e) {
+            e.printStackTrace();
+            fail();
+        }
+        setupMock(httpClient, 201, payment.toString());
 
         ret = httpClient.loadInBackground();
 
@@ -63,6 +76,7 @@ public class HTTPClientTest {
             assertEquals(expectedPayment, actualPayment);
         } catch (JSONException e) {
             e.printStackTrace();
+            fail();
         }
     }
 
@@ -70,8 +84,18 @@ public class HTTPClientTest {
     public void testRegisterPayment_NG_categoryIsAbsent() {
         String[] inputs = {"2015-01-01", "for http client test", null, "100", "expense"};
 
-        HTTPClient httpClient = new HTTPClient(getContext(), inputs);
-        setupMock(httpClient, 400, "[{\"error_code\":\"absent_param_category\"}]");
+        HTTPClient httpClient = new HTTPClient(getContext());
+        httpClient.createPayment(inputs);
+        JSONArray errors = new JSONArray();
+        try {
+            JSONObject error = new JSONObject();
+            error.put("error_code", "absent_param_category");
+            errors.put(error);
+        } catch (JSONException e) {
+            e.printStackTrace();
+            fail();
+        }
+        setupMock(httpClient, 400, errors.toString());
 
         ret = httpClient.loadInBackground();
 
@@ -83,8 +107,18 @@ public class HTTPClientTest {
     public void testRegisterPayment_NG_dateIsInvalid() {
         String[] inputs = {"invalid_date", "テスト用データ", "テスト", "100", "expense"};
 
-        HTTPClient httpClient = new HTTPClient(getContext(), inputs);
-        setupMock(httpClient, 400, "[{\"error_code\":\"invalid_param_date\"}]");
+        HTTPClient httpClient = new HTTPClient(getContext());
+        httpClient.createPayment(inputs);
+        JSONArray errors = new JSONArray();
+        try {
+            JSONObject error = new JSONObject();
+            error.put("error_code", "invalid_param_date");
+            errors.put(error);
+        } catch (JSONException e) {
+            e.printStackTrace();
+            fail();
+        }
+        setupMock(httpClient, 400, errors.toString());
 
         ret = httpClient.loadInBackground();
 
@@ -96,8 +130,21 @@ public class HTTPClientTest {
     public void testRegisterPayment_NG_multipleInputsIsAbsent() {
         String[] inputs = {"2015-01-01", null, "テスト", null, "expense"};
 
-        HTTPClient httpClient = new HTTPClient(getContext(), inputs);
-        setupMock(httpClient, 400, "[{\"error_code\":\"absent_param_content\"}, {\"error_code\":\"absent_param_price\"}]");
+        HTTPClient httpClient = new HTTPClient(getContext());
+        httpClient.createPayment(inputs);
+        JSONArray errors = new JSONArray();
+        try {
+            JSONObject error = new JSONObject();
+            error.put("error_code", "absent_param_content");
+            errors.put(error);
+            error = new JSONObject();
+            error.put("error_code", "absent_param_price");
+            errors.put(error);
+        } catch (JSONException e) {
+            e.printStackTrace();
+            fail();
+        }
+        setupMock(httpClient, 400, errors.toString());
 
         ret = httpClient.loadInBackground();
 
@@ -109,8 +156,21 @@ public class HTTPClientTest {
     public void testRegisterPayment_NG_multipleInputsIsInvalid() {
         String[] inputs = {"invalid_date", "テスト用データ", "テスト", "-100", "expense"};
 
-        HTTPClient httpClient = new HTTPClient(getContext(), inputs);
-        setupMock(httpClient, 400, "[{\"error_code\":\"invalid_param_date\"}, {\"error_code\":\"invalid_param_price\"}]");
+        HTTPClient httpClient = new HTTPClient(getContext());
+        httpClient.createPayment(inputs);
+        JSONArray errors = new JSONArray();
+        try {
+            JSONObject error = new JSONObject();
+            error.put("error_code", "invalid_param_date");
+            errors.put(error);
+            error = new JSONObject();
+            error.put("error_code", "invalid_param_price");
+            errors.put(error);
+        } catch (JSONException e) {
+            e.printStackTrace();
+            fail();
+        }
+        setupMock(httpClient, 400, errors.toString());
 
         ret = httpClient.loadInBackground();
 
@@ -121,25 +181,60 @@ public class HTTPClientTest {
     @Test
     public void testGetSettlement_OK() {
         HTTPClient httpClient = new HTTPClient(getContext());
-        setupMock(httpClient, 200, "[{\"2000-01\":\"1000\"}, {\"2000-02\":\"-100\"}]");
+        httpClient.getSettlements();
+        JSONArray settlements = new JSONArray();
+        try {
+            JSONObject settlement = new JSONObject();
+            settlement.put("date", "2000-01");
+            settlement.put("price", "1000");
+            settlements.put(settlement);
+            settlement = new JSONObject();
+            settlement.put("date", "2000-02");
+            settlement.put("price", "-100");
+            settlements.put(settlement);
+        } catch (JSONException e) {
+            e.printStackTrace();
+            fail();
+        }
+        setupMock(httpClient, 200, settlements.toString());
 
         ret = httpClient.loadInBackground();
 
         assertStatusCode(200);
 
         try {
-            JSONObject body = new JSONObject(ret.get("body").toString());
-            assertEquals("1000", body.getString("2000-01"));
-            assertEquals("-100", body.getString("2000-02"));
+            JSONArray body = new JSONArray(ret.get("body").toString());
+            assertEquals("2000-01", body.getJSONObject(0).get("date"));
+            assertEquals("1000", body.getJSONObject(0).get("price"));
+            assertEquals("2000-02", body.getJSONObject(1).get("date"));
+            assertEquals("-100", body.getJSONObject(1).get("price"));
         } catch (JSONException e) {
             e.printStackTrace();
+            fail();
         }
     }
 
     @Test
     public void testGetCategories_OK() {
-        HTTPClient httpClient = new HTTPClient(getContext(), "");
-        setupMock(httpClient, 200, "[{\"id\":\"1\", \"name\":\"test\", \"description\":\"test category\"}, {\"id\":\"2\", \"name\":\"test2\", \"description\":\"test category\"}]");
+        HTTPClient httpClient = new HTTPClient(getContext());
+        httpClient.getCategories("");
+        JSONArray categories = new JSONArray();
+        try {
+            JSONObject category = new JSONObject();
+            category.put("id", "1");
+            category.put("name", "test");
+            category.put("description", "test category");
+            categories.put(category);
+            category = new JSONObject();
+            category.put("id", "2");
+            category.put("name", "test2");
+            category.put("description", "test category");
+            categories.put(category);
+        } catch (JSONException e) {
+            e.printStackTrace();
+            fail();
+        }
+        setupMock(httpClient, 200, categories.toString());
 
         ret = httpClient.loadInBackground();
 
@@ -155,6 +250,55 @@ public class HTTPClientTest {
             assertEquals("test category", body.getJSONObject(1).getString("description"));
         } catch (JSONException e) {
             e.printStackTrace();
+            fail();
+        }
+    }
+
+    @Test
+    public void testGetPayments_OK() {
+        HTTPClient httpClient = new HTTPClient(getContext());
+        httpClient.getPayments(new HashMap<String, String>());
+        JSONArray payments = new JSONArray();
+        try {
+            JSONObject payment = new JSONObject();
+            payment.put("id", "1");
+            payment.put("payment_type", "income");
+            payment.put("date", "1000-01-01");
+            payment.put("content", "test");
+            JSONArray categories = new JSONArray();
+            JSONObject category = new JSONObject();
+            category.put("id", "1");
+            category.put("name", "test");
+            category.put("description", "test category");
+            categories.put(category);
+            payment.put("categories", categories);
+            payment.put("price", "1000");
+            payments.put(payment);
+        } catch (JSONException e) {
+            e.printStackTrace();
+            fail();
+        }
+        setupMock(httpClient, 200, payments.toString());
+
+        ret = httpClient.loadInBackground();
+
+        assertStatusCode(200);
+
+        try {
+            JSONArray body = new JSONArray(ret.get("body").toString());
+            assertEquals("1", body.getJSONObject(0).getString("id"));
+            assertEquals("income", body.getJSONObject(0).getString("payment_type"));
+            assertEquals("1000-01-01", body.getJSONObject(0).getString("date"));
+            assertEquals("test", body.getJSONObject(0).getString("content"));
+            assertEquals("1000", body.getJSONObject(0).getString("price"));
+
+            JSONArray categories = body.getJSONObject(0).getJSONArray("categories");
+            assertEquals("1", categories.getJSONObject(0).getString("id"));
+            assertEquals("test", categories.getJSONObject(0).getString("name"));
+            assertEquals("test category", categories.getJSONObject(0).getString("description"));
+        } catch (JSONException e) {
+            e.printStackTrace();
+            fail();
         }
     }
 
@@ -185,11 +329,13 @@ public class HTTPClientTest {
             assertEquals(expectedErrors, actualErrors);
         } catch (JSONException e) {
             e.printStackTrace();
+            fail();
         }
     }
 
     private void setupMock(HTTPClient httpClient, int statusCode, final String responseBody) {
         try {
+            when(con.getRequestMethod()).thenReturn("");
             when(con.getResponseCode()).thenReturn(statusCode);
             when(con.getInputStream()).thenReturn(new InputStream() {
                 private int position = 0;